contextcore-coyote/pyproject.toml
[project]
name = "contextcore-coyote"
version = "0.1.0"
description = "Coyote (Wiisagi-ma'iingan) - Multi-agent incident resolution pipeline for ContextCore"
readme = "README.md"
license = {text = "Equitable Use License v1.0"}
requires-python = ">=3.12"
authors = [
    {name = "ContextCore Contributors"}
]
keywords = [
    "contextcore",
    "incident-response",
    "multi-agent",
    "pipeline",
    "observability",
    "automation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Monitoring",
    "Topic :: Software Development :: Bug Tracking",
]

dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "httpx>=0.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
contextcore = ["contextcore>=0.1.0"]
llm = ["anthropic>=0.18"]
github = ["pygithub>=2.0"]
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]
all = [
    "contextcore>=0.1.0",
    "anthropic>=0.18",
    "pygithub>=2.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "ruff>=0.1",
    "black>=23.0",
]

[project.urls]
Homepage = "https://github.com/contextcore/contextcore-coyote"
Documentation = "https://github.com/contextcore/contextcore-coyote#readme"
Repository = "https://github.com/contextcore/contextcore-coyote"

[project.scripts]
contextcore-coyote = "contextcore_coyote.cli:main"
coyote = "contextcore_coyote.cli:main"
pup = "contextcore_coyote.pup_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/contextcore_coyote"]

[tool.hatch.build]
include = [
    "src/contextcore_coyote/**/*.py",
    "src/contextcore_coyote/**/*.md",
    "src/contextcore_coyote/**/*.yaml",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
